Fix bug where language parser would return incorrect name
Originally, the parser would check if the string merely *contains*
the language URL path, however this would have lead to issues for
language IDs of different locales (e.g. 'spa' and 'spa_es'), since
'spa' is contained within 'spa_es'. To fix this, the parser now
checks if the URL ends with the language URL path, ensuring that
the language ID matches exactly.