Use Alfred-consistent return symbol
Even though ⏎ is technically the symbol for "Return" per the Unicode
standard, Alfred uses ↩ in its interface to represent pressing the
Enter/Return key on a result. This is perhaps also why the latter symbol
is used in the new ChatGPT/DALL-E workflow. Therefore, we should use the
same symbol for greater consistency with the convention that Alfred has
already set (regardless of the technical correctness of the semantics).